delete stale chassis from sbdb when its chassis node annotation changes
Today, when a node is deleted from the cluster, its chassis entry is
deleted from sbdb by ovnkube-master, but it can be soon added back by
ovn-controller. This stale chassis entry prevents new one being added to
sbdb when the node is added back to the cluster with a different
chassis id.

This commit checks node's chassis annotation and delete the stale
chassis from sbdb if the annotation suggests there is a change of
chassis id.

Signed-off-by: Yun Zhou <yunz@nvidia.com>