fix potential pod IP leak in case that addLogicalPortToNetwork() fails
when addLogicalPortToNetwork() fails after it successfully allocated pod
IPs and updated the pod annotation, there is a case new pod IPs can be
allocated to the pod during retry if the informer cache does not cache
up with the latest pod annotation update. This would cause leak of the
old IPs allocated to the pod.

Signed-off-by: Yun Zhou <yunz@nvidia.com>